requirements.txt
venusian==1.0a8
wheel==0.23.0
zope.dottedname==4.0.1
zope.interface==4.1.1
